chore: drop custom css variable parser
