+++ tests/helpers/meditationContrast.test.js
@@ -3,11 +3,20 @@ import { describe, it, expect } from "vitest";
 import { readFileSync } from "fs";
 import { resolve } from "path";
 import { hex } from "wcag-contrast";
-import { parseCssVariables } from "../../src/helpers/cssVariableParser.js";
+import postcss from "postcss";
 
 function getBaseVars() {
   const css = readFileSync(resolve("src/styles/base.css"), "utf8");
-  return parseCssVariables(css);
+  const vars = {};
+  const root = postcss.parse(css);
+  root.walkRules(":root", (rule) => {
+    rule.walkDecls((decl) => {
+      if (decl.prop.startsWith("--")) {
+        vars[decl.prop] = decl.value.trim();
+      }
+    });
+  });
+  return vars;
 }
 
 describe("meditation quote color contrast", () => {
